+++ wsdr/src/dsp/cbb_main.h
@@ -9,7 +9,7 @@ rf_decimator *cbb_rf_decimator();
 
 int cbb_new_spectrum_available();
 
-int cbb_get_spectrum_payload(char *buf, int buf_len, int spectrum_gain_db);
+int cbb_get_spectrum_payload(char *buf, int buf_len, double spectrum_gain_db);
 
 void cbb_close();
